Make pdal with embedded boost usable for external linking
Linking against pdal which has been compiled with embedded boost
requires the pdalboost headers to be installed, and to link against the
associated libpdalboost library.

To make this work with both shared and static versions of libpdal, we
build the required parts of boost directly into libpdal and install the
required headers.  This also avoids the need to manually add -fPIC in
the cmake script.